separated asyncWorker into its own file
